+++ NodeServer/app.js
@@ -1,0 +1,21 @@
+const express = require('express');
+const graphqlHTTP = require('express-graphql');
+const schema = require('./schema/schema');
+const mongoose = require('mongoose');
+
+const app = express();
+
+mongoose.connect("mongodb+srv://dbUser:dbuserpass@database-pkdd1.azure.mongodb.net/test?retryWrites=true&w=majority", { useNewUrlParser: true });
+mongoose.connection.once("open", () => {
+    console.log("Connected to MongoDB");
+})
+
+
+app.use('/graphql', graphqlHTTP({
+    schema,
+    graphiql: true,
+}));
+
+app.listen(4001, () => {
+    console.log('now listening for requests on port 4001');
+});
